requirements.txt
appdirs==1.4.3
chardet==2.3.0
Django==1.10.6
django-debug-toolbar
djangorestframework==3.3.3
djangorestframework-jwt==1.7.2
djangorestframework-bulk==0.2.1
MySQL-python==1.2.5
packaging==16.8
pyparsing==2.2.0
six==1.10.0
unicodecsv==0.14.1
